test: rename and add TestConfig to TaxControllerIntegrationTest
